GenomicOffsets.jl Documentation

The GenomicOffsets Julia package implements efficient versions of some of the most popular Genomic offset metrics.

Disclaimer: this documentation is not finished and just provides a quick overview of the syntax.

Getting started

The package provides a miminal dataset you can use. You can load it as follos.

using GenomicOffsets
Y, X, Xpred = data;

([0 0 … 0 0; 0 0 … 0 0; … ; 0 0 … 0 0; 1 0 … 2 0], [0.553275 0.810589 -0.5021923505314568 -0.332923350974514; 0.782761 0.113466 0.13153116532730252 1.3631137069224968; … ; 0.892056 0.579782 -1.178683140724628 -1.0854529069457097; 0.648891 0.378069 -1.1740347584692243 0.5769373005244982], [0.426251 0.826528 0.02817176621690603 -1.8554452269695023; 0.590292 -0.120721 -0.35670340577785264 -1.0185647149432182; … ; 0.76328 0.619325 0.8140321710708628 -1.1708244066000315; 0.469984 0.200247 -1.3200243959749645 -1.1390516111210724])

On the one hand, Y is the genotype matrix (or allele frequencies) of the current locally adapted individuals (or populations). We have 100 individuals and 500 loci.

size(Y)

(100, 5261)

On the other hand, X is the current environmental matrix. We have 100 individuals and 2 ecological predictors. Xpred are the same individuals (although this is not mandatory) for which we have a forecasted ecological predictors.

@assert size(X) == size(Xpred)
size(X)

(100, 4)

We can compute the RONA (and the rest of supported methods) by first, fitting the model and then computing the predicted genomic offset.

rona = fit(RONA, Y, X)
genomic_offset(rona, X, Xpred)

100×1 Matrix{Float64}:
 0.04930412355509537
 0.09685110436545971
 0.031955998579532574
 0.057903936496619686
 0.06747793104523755
 0.07838283024137747
 0.034071076534702927
 0.05671776002159728
 0.11209514598766128
 0.050635672102255044
 ⋮
 0.05847056037198923
 0.044459426014584816
 0.1000681402606062
 0.076366019095105
 0.0779261159151334
 0.09861395620054535
 0.06261235895467696
 0.05449853921861244
 0.07504407351624044

Other genomic offset metrics have additional parameters. For example, if computing the geometric genomic offset, you may want to provide the number of latent factors of the LFMM model (and not use the Tracy-Widom test to determined it).

geometric = fit(GeometricGO, Y, X, 3)
genomic_offset(geometric, X, Xpred)

100-element Vector{Float64}:
 0.00593633515140897
 0.023044843347597432
 0.002627190938054718
 0.009020228830284487
 0.010936690454746233
 0.017951498051068514
 0.003197467471179165
 0.008461430697171007
 0.026531050213325375
 0.007267500338404611
 ⋮
 0.007924421867094151
 0.0057941693706635
 0.028160635512418916
 0.012680311728430321
 0.014034542738334124
 0.020945759860081285
 0.010466825157368317
 0.006955130907343611
 0.014350598499661762

To see a more complicated use case, let's compute the 95% confidence niterval using bootstrapping. We do it by computing the geometric genomic offset repeatly with different loci sampled with replacement. At every iteration, (a) the number of latent factors will be determined using the TracyWidom test, (b) a set of putatively adaptative loci will be identified with an F-test (corrected using genomic control) based on the LFMM model and (c) the geometric genomic offste is computed.

using Random, StatsBase
nboots = 100
offsets = bootstrap_with_candidates(GeometricGO, Y, X, Xpred, nboots;candidates_threshold=0.05)
@assert size(offsets) == (size(Y, 1), nboots)
confint95 = [quantile(ind, [0.025, 0.975]) for ind in eachrow(offsets)]

100-element Vector{Vector{Float64}}:
 [0.023119455700177184, 0.0321081353245525]
 [0.10672556311309765, 0.15183612433141028]
 [0.010624860918386821, 0.015076507773363957]
 [0.03958325966262796, 0.057458046917971524]
 [0.044485644082217925, 0.06217204009706437]
 [0.09502106762794617, 0.13455711336624934]
 [0.014724977103774423, 0.020931993880297182]
 [0.03587750589904476, 0.05049396049330904]
 [0.09206283983794634, 0.12567920267863691]
 [0.035613297496607285, 0.051904918757109696]
 ⋮
 [0.020772036547795907, 0.030939270850262855]
 [0.028804463056015688, 0.043650300999771054]
 [0.14172138498254916, 0.20768252971938583]
 [0.04145362950473591, 0.06613990515931592]
 [0.049722496616081795, 0.07258601668515333]
 [0.05651186034820065, 0.08154955531258602]
 [0.048813534626198835, 0.07485408033839958]
 [0.028165325827914315, 0.04161367139971086]
 [0.06877872912384338, 0.09816115225361913]

References

GenomicOffsets.LFMM_Ftest — Method

LFMM_Ftest(model::LFMM, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}; genomic_control::Bool=true, center=true) where {T1<:Real,T2<:Real}

Compute F-Statistic (F-test) for the Linear Fixed Effects Mixed Model (LFMM). TODO: Add reference & description.

Arguments

model: A LFMM{T<:Real} data structure (obtained from RidgeLFMM).

Y: A centered genotype matrix of size NxL.

X: A centered environmental matrix of size NxP.

genomic_control: If true, apply genomic control to the F-statistic.

center: If true, both the genotype matrix and the environmental matrix are centered. If false, both matrices are assumed to be centered.

Returns

A vector of p-values for loci in the genotype matrix.

GenomicOffsets.RidgeLFMM — Method

RidgeLFMM(Y::Matrix{T}, X::Matrix{T}, K::Int, λ::Float64) where T<:Real

Ridge solutions for the Linear Fixed Effects Mixed Model (LFMM) as described by Caye et al. (2019).

\[Y = X B^T + W = X B^T + U V^T\]

Arguments

Y: A centered genotype matrix of size NxL.

X: Environmental matrix of size NxP.

K: Number of latent factors.

λ: Regularization parameter (by 1e-5)

center: If true, both the genotype matrix and the environmental matrix are centered. If false, both matrices are assumed to be centered.

Returns

A LFMM{T<:Real} data structure with the latent factors U, Vt, and the effect sizes Bt.

GenomicOffsets.TracyWidom — Method

TracyWidom(eigenvalues::AbstractVector{T}; sort_eigenvalues::Bool=true) where T<:Real

Compute the Tracy-Widom statistics and p-values for a given set of eigenvalues. TODO: add reference.

Arguments

eigenvalues::AbstractVector{T}: A vector of eigenvalues.

sort_eigenvalues::Bool=true: If true, sort the eigenvalues in descending order. If false, the eigenvalues are assumed to be sorted in descending order.

Returns

A tuple with the Tracy-Widom statistics and p-values. For a given threshold, it is advised not to keep all eigenvalues below the threshold, but up to the first one that is above the threshold (not included).

GenomicOffsets.bootstrap_with_candidates — Method

bootstrap_with_candidates(::Type{GeometricGO}, rng::Random.AbstractRNG,Y::AbstractMatrix{T1}, X::AbstractMatrix{T2},Xpred::AbstractMatrix{T2}, nboot::Int=500; candidates_threshold::Real=0.05, genomic_control::Bool=true, tw_threshold::Real=0.001) where {T1<:Real,T2<:Real}

Compute the Geometric genomic offset using a bootstrap approach. For every, bootstrap iteration, the model is fitted using a random subset of the columns of Y. The Tracy-Widom test is used to estimate the number of latent factors. The F-test is used to select putatively adaptative loci. The genomic offset is computed for the selected loci.

Arguments

rng::Random.AbstractRNG: Random number generator. If not provided, the global RNG is used.

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T2}: NxP matrix with environmental data.

Xpred::AbstractMatrix{T2}: NxP matrix with the predicted / altered environmental matrix.

nboot::Int=500: Number of bootstrap iterations.

Optional arguments:

candidates_threshold::Real=0.05: Threshold for the F-test. Better to be permisive.

genomic_control::Bool=true: Apply genomic control to the F-test.

tw_threshold::Real=0.001: Threshold for the Tracy-Widom test. Better to be conservative (as the number of latent factors is often overconservative).

Returns

A matrix of size NxNboot with the genomic offset values. If no candidate loci are found, the row is filled with zeros.

GenomicOffsets.bootstrap_with_candidates — Method

bootstrap_with_candidates(::Type{GradientForestGO}, rng::Random.AbstractRNG,Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}, Xpred::AbstractMatrix{T2}, nboot::Int=500; ntrees::Int=100, candidates_threshold::Real=0.05, genomic_control::Bool=true, tw_threshold::Real=0.001) where {T1<:Real,T2<:Real}

Compute the Gradient Forest genomic offset using a bootstrap approach. For every, bootstrap iteration, the model is fitted using a random subset of the columns of Y. The Tracy-Widom test is used to estimate the number of latent factors. The F-test is used to select putatively adaptative loci. The genomic offset is computed for the selected loci.

Arguments

rng::Random.AbstractRNG: Random number generator. If not provided, the global RNG is used.

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T2}: NxP matrix with environmental data.

Xpred::AbstractMatrix{T2}: NxP matrix with the predicted / altered environmental matrix.

nboot::Int=500: Number of bootstrap iterations.

Optional arguments:

ntrees::Int=100: Number of trees each forest (one per allele) will have.

candidates_threshold::Real=0.05: Threshold for the F-test. Better to be permisive.

genomic_control::Bool=true: Apply genomic control to the F-test.

tw_threshold::Real=0.001: Threshold for the Tracy-Widom test. Better to be conservative (as the number of latent factors is often overconservative).

Returns

A matrix of size NxNboot with the genomic offset values. If no candidate loci are found, the row is filled with zeros.

GenomicOffsets.bootstrap_with_candidates — Method

bootstrap_with_candidates(::Type{RDAGO}, rng::Random.AbstractRNG, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}, Xpred::AbstractMatrix{T2}, nboot::Int=500; candidates_threshold::Real=0.05, genomic_control::Bool=true, tw_threshold::Real=0.001) where {T1<:Real, T2<:Real}

Compute the RDA genomic offset using a bootstrap approach. For every, bootstrap iteration, the model is fitted using a random subset of the columns of Y. The Tracy-Widom test is used to estimate the number of latent factors. The F-test is used to select putatively adaptative loci. The genomic offset is computed for the selected loci.

Arguments

rng::Random.AbstractRNG: Random number generator. If not provided, the global RNG is used.

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T2}: NxP matrix with environmental data.

Xpred::AbstractMatrix{T2}: NxP matrix with the predicted / altered environmental matrix.

nboot::Int=500: Number of bootstrap iterations.

Optional arguments:

candidates_threshold::Real=0.05: Threshold for the F-test. Better to be permisive.

genomic_control::Bool=true: Apply genomic control to the F-test.

tw_threshold::Real=0.001: Threshold for the Tracy-Widom test. Better to be conservative (as the number of latent factors is often overconservative).

Returns

A matrix of size NxNboot with the genomic offset values. If no candidate loci are found, the row is filled with zeros.

GenomicOffsets.bootstrap_with_candidates — Method

bootstrap_with_candidates(::Type{RONA}, rng::Random.AbstractRNG, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}, Xpred::AbstractMatrix{T2}, nboot::Int=500; candidates_threshold::Real=0.05, genomic_control::Bool=true, tw_threshold::Real=0.001) where {T1<:Real, T2<:Real}

Compute the genomic offset for the RONA model using a bootstrap approach. For every, bootstrap iteration, the model is fitted using a random subset of the columns of Y. The Tracy-Widom test is used to estimate the number of latent factors. The F-test is used to select putatively adaptative loci. The genomic offset is computed for the selected loci.

Arguments

rng::Random.AbstractRNG: Random number generator. If not provided, the global RNG is used.

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T2}: NxP matrix with environmental data.

Xpred::AbstractMatrix{T2}: NxP matrix with the predicted / altered environmental matrix.

nboot::Int=500: Number of bootstrap iterations.

Optional arguments:

candidates_threshold::Real=0.05: Threshold for the F-test. Better to be permisive.

genomic_control::Bool=true: Apply genomic control to the F-test.

tw_threshold::Real=0.001: Threshold for the Tracy-Widom test. Better to be conservative (as the number of latent factors is often overconservative).

Returns

A matrix of size NxNboot with the genomic offset values. If no candidate loci are found, the row is filled with zeros.

GenomicOffsets.fit — Method

fit(::Type{GeometricGO}, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}) where {T1<:Real,T2<:Real}

Fit the Geometric genomic offset model (that is, fit a LFMM) using the data Y and X.

Arguments

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T}: NxP matrix with environmental data.

K: Number of latent factors to use when fitting th e LFMM model. It can be determined using the elbow rule of thumb. If nothing is provided, the Tracy-Widom statistic to determine the number of latent factors (which might be overconservative).

Optional arguments:

center::Bool=true: Center both the genotype and environmental data. Predicted environmental data is centered using the same mean as X. If false, data is assumed to be centered.

scale::Bool=true: Scale the environmental data. Predicted environmental data is scaled using the same standard deviation as X. If false, data is assumed to be scaled.

tw_threshold::Real=1e-3: Threshold to determine the number of latent factors if using the Tracy-Widom statistic. By default, it uses 1e-3.

λ::Real=1e-5: Regularization parameter for the LFMM.

Returns

A GeometricGO model.

GenomicOffsets.fit — Method

fit(::Type{GradientForestGO}, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}; ntrees::Int=100) where {T1<:Real,T2<:Real}

Fit the Gradient forest genomic offset model (that is, fit a Gradient Forest) using the data Y and X.

Arguments

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T}: NxP matrix with environmental data.

ntrees: Number of trees each forest (one per allele) will have.

Returns

A GradientForestGO model.

GenomicOffsets.fit — Method

fit(::Type{RDAGO}, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}) where {T1<:Real,T2<:Real}

Fit the RDA model (that is, apply Redundancy Analysis) using the data Y and X.

Arguments

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T}: NxP matrix with environmental data.

Returns

A RDAGO model.

GenomicOffsets.fit — Method

fit(::Type{RONA}, Y::AbstractMatrix{T1}, X::AbstractMatrix{T2}) where {T1<:Real,T2<:Real}

Fit the RONA model (that is, solve the least squares problem) using the data Y and X.

Arguments

Y::AbstractMatrix{T1}: NxL matrix with individuals genotypes or allele frequencies.

X::AbstractMatrix{T}: NxP matrix with environmental data.

Returns

A RONA model.

GenomicOffsets.genomic_offset — Method

genomic_offset(model::GeometricGO, X::AbstractMatrix{T}, Xpred::AbstractMatrix{T}) where T<:Real

Compute the Geometric genomic offset.

Arguments

model::GeometricGO: A GeometricGO model.

X::AbstractMatrix{T}: NxP matrix with environmental data. Data will be scaled and centered if the model was fitted with center and scale set to true.

Xpred::Matrix{T}: NxP matrix with the predicted / altered environmental matrix. Data will be scaled and centered if the model was fitted with center and scale set to true.

candidates::AbstractVector{Integer}: Indices of the candidate loci.

Returns

A Vector{Float64} of length N with the Geometric genomic offset values.

GenomicOffsets.genomic_offset — Method

genomic_offset(model::GeometricGO, X::AbstractMatrix{T}, Xpred::AbstractMatrix{T}) where T<:Real

Compute the Geometric genomic offset.

Arguments

model::GeometricGO: A GeometricGO model.

X::AbstractMatrix{T}: NxP matrix with environmental data. Data will be scaled and centered if the model was fitted with center and scale set to true.

Xpred::Matrix{T}: NxP matrix with the predicted / altered environmental matrix. Data will be scaled and centered if the model was fitted with center and scale set to true.

Returns

A Vector{Float64} of length N with the Geometric genomic offset values.

GenomicOffsets.genomic_offset — Method

genomic_offset(model::GradientForestGO, X::AbstractMatrix{T}, Xpred::AbstractMatrix{T}) where T<:Real

Compute the Gradient Forest genomic offset. Attention! Our implementation does not perform extrapolation, which means that you may obtain unpredictable results if you try so.

Arguments

model::GradientForestGO: A GeometricGO model.

X::AbstractMatrix{T}: NxP matrix with environmental data. Data will be scaled and centered if the model was fitted with center and scale set to true.

Xpred::Matrix{T}: NxP matrix with the predicted / altered environmental matrix. Data will be scaled and centered if the model was fitted with center and scale set to true.

Returns

A Vector{Float64} of length N with the Gradient forest genomic offset values.

GenomicOffsets.genomic_offset — Method

genomic_offset(model::RDAGO, X::AbstractMatrix{T}, Xpred::AbstractMatrix{T}) where T<:Real

Compute the RDA genomic offset.

Arguments

model::RDAGO: A RDAGO model.

X::AbstractMatrix{T}: NxP matrix with environmental data.

Xpred::Matrix{T}: NxP matrix with the predicted / altered environmental matrix.

Returns

A Vector{Float64} of length N with the RDAGO genomic offset values.

GenomicOffsets.genomic_offset — Method

genomic_offset(model::RONA, X::AbstractMatrix{T}, Xpred::AbstractMatrix{T}) where T<:Real

Compute the genomic offset for the RONA model.

Arguments

model::RONA: A RONA model.

X::AbstractMatrix{T}: NxP matrix with environmental data.

Xpred::Matrix{T}: NxP matrix with the predicted / altered environmental matrix.

Returns

A Vector{Float64} of length N with the RONA genomic offset values.

GenomicOffsets.GeometricGO — Type

Geometric genomic offset. TODO: Add reference

GenomicOffsets.GradientForestGO — Type

Gradient genomic offset. TODO: Add reference

GenomicOffsets.LFMM — Type

LFMM{T<:Real}

A struct to store the results of the Linear Fixed Effects Mixed Model (LFMM) as described by Caye et al. (2019).

\[Y = X B^T + W = X B^T + U V^T\]

GenomicOffsets.RDAGO — Type

Redundancy Analysis (RDA) genomic offset. TODO: Add reference

GenomicOffsets.RONA — Type

Risk of non-adaptnesss (RONA) genomic offset. First proposed by Rellstab et al. (2016).